Exponential back-off support for megadl
This will attempt to download a file 5 times, only if the error is an
HTTP_ERROR. In other words, if the error was not a download error, then
it will be treated as normal, and the application will exit. The maximum
attempts are hard-coded to 5, but it can easily be changed to allow the
user to control this value. The wait time doubles on each attempt
so, e.g. 2 seconds, 4, 8, 16, etc.